+++ src/permifrost/core/permissions/utils/snowflake_connector.py
@@ -193,6 +193,16 @@ class SnowflakeConnector:
 
         return roles
 
+    def get_current_user(self) -> str:
+        query = f"SELECT CURRENT_USER() AS USER"
+        result = self.run_query(query).fetchone()
+        return result["user"].lower()
+
+    def get_current_role(self) -> str:
+        query = f"SELECT CURRENT_ROLE() AS ROLE"
+        result = self.run_query(query).fetchone()
+        return result["role"].lower()
+
     def run_query(self, query: str):
 
         with self.engine.connect() as connection: