+++ tests/permifrost/core/permissions/test_snowflake_spec_loader.py
@@ -1,6 +1,7 @@
 import pytest
 import os
 
+from permifrost.core.permissions import SpecLoadingError
 from permifrost.core.permissions.snowflake_spec_loader import SnowflakeSpecLoader
 from typing import Dict, List, Any
 
@@ -35,6 +36,12 @@ class MockSnowflakeConnector:
 
     def show_roles_granted_to_user(self, user) -> List[str]:
         return []
+
+    def get_current_user(self) -> str:
+        return ""
+
+    def get_current_role(self) -> str:
+        return "securityadmin"
 
 
 @pytest.fixture
@@ -103,3 +110,36 @@ class TestSnowflakeSpecLoader:
             os.path.join(test_dir, "specs", "snowflake_spec_blank.yml"), mock_connector
         )
         mock_connector.show_users.assert_not_called()
+
+    def test_check_permissions_on_snowflake_server_as_securityadmin(
+        self, test_dir, mocker, mock_connector
+    ):
+        mocker.patch.object(
+            MockSnowflakeConnector, "get_current_role", return_value="securityadmin"
+        )
+        SnowflakeSpecLoader(
+            os.path.join(test_dir, "specs", "snowflake_spec_blank.yml"), mock_connector
+        )
+        mock_connector.get_current_role.assert_called()
+
+    def test_check_permissions_on_snowflake_server_not_as_securityadmin(
+        self, test_dir, mocker, mock_connector
+    ):
+        mocker.patch.object(
+            MockSnowflakeConnector, "get_current_role", return_value="notsecurityadmin"
+        )
+        with pytest.raises(SpecLoadingError) as context:
+            SnowflakeSpecLoader(
+                os.path.join(test_dir, "specs", "snowflake_spec_blank.yml"),
+                mock_connector,
+            )
+            mock_connector.get_current_role.assert_called()
+
+    def test_check_permissions_on_snowflake_server_gets_current_user_info(
+        self, test_dir, mocker, mock_connector
+    ):
+        mocker.patch.object(MockSnowflakeConnector, "get_current_user")
+        SnowflakeSpecLoader(
+            os.path.join(test_dir, "specs", "snowflake_spec_blank.yml"), mock_connector
+        )
+        mock_connector.get_current_user.assert_called()